stil a lot more to do

+++ libft/strtoint/strtoint.h
@@ -6,7 +6,7 @@
 /*   By: aghergut <aghergut@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 17:35:19 by aghergut          #+#    #+#             */
-/*   Updated: 2024/12/21 18:06:22 by aghergut         ###   ########.fr       */
+/*   Updated: 2025/02/11 18:31:06 by aghergut         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,11 @@
 # define STRTOINT_H
 
 # include "../strings/strings.h"
+# include <limits.h>
 
 //int to str
 char	*ft_itoa(int n);
 //str to int
-int		ft_atoi(const char *str);
+long	ft_atoi(const char *str);
 
 #endif
